Move BigInt type to a separate module
